Two New Maps created and issues fixed
Two Open World maps created, one with dynamic and static cubes and spheres, also has static ramps. The other map is a football pitch to score goals with a dynamic sphere. A pause menu has been created for the football map.

The deconstructors for components were never called, this has been fixed. Fixing this caused an issue with closing the game when after returning to the main menu, the physx extensions need closing but causes a crash, so checking if the physx foundation is null or not fixes this issue